Add crafting recipes (Closes #94)

+++ src/content/machines/crafter.rs
@@ -16,7 +16,7 @@ pub struct Crafter {
 
     /// Crafting time left
     /// `None` if nothing is currently being crafting
-    crafting_time_left: Option<u8>,
+    crafting_time_left: Option<u16>,
 }
 
 impl Crafter {
